Avoid AttributeError if calling stop before start
Remove the related pylint disable

+++ tests/client-dbus/tests/udev/_utils.py
@@ -325,6 +325,9 @@ class _Service:
     Start and stop stratisd.
     """
 
+    def __init__(self):
+        self._service = None
+
     def start_service(self):
         """
         Starts the stratisd service if it is not already started. Verifies
@@ -361,7 +364,7 @@ class _Service:
                 "No D-Bus interface for stratisd found although stratisd appears to be running"
             ) from err
 
-        self._service = service  # pylint: disable=attribute-defined-outside-init
+        self._service = service
         return self
 
     def stop_service(self):
@@ -369,6 +372,9 @@ class _Service:
         Stops the stratisd daemon previously spawned.
         :return: None
         """
+        if self._service is None:
+            return
+
         self._service.send_signal(signal.SIGINT)
         self._service.wait(timeout=30)
         if next(processes("stratisd"), None) is not None:
